Added the Image Slideshow/Management Code
This code contains the main functionality for Image Slideshow.
Involves creating a metadata JSON file that contains an Image's information: Their file path, as well as their expiration date.
There is also a section of the code where a change is needed when we are set and done. This change is needed because the code's functionality is based on a direct folder path rather than a pull from some sort of database where I would assume images would be stored.